install_translations.sh: can install from directory, not only tarball
BUT currently unused

to be used once the Rust code works and we get to a proper installation part

+++ service/package/install_translations.sh
@@ -1,10 +1,32 @@
 #! /bin/sh
 
-# a helper script for compiling and installing the translations
+# A helper script for compiling and installing the translations
+# Usage:
+#   install_translations po_archive.tar.bz2
+#   install_translations -d po_dir
+# Where the contents are ll.po ll_TT.po
 
-PODIR=$(mktemp --directory --suffix "-agama-po")
+set -eu
 
-tar xfjv "$1" -C "$PODIR"
-find "$PODIR" -name "*.po" -exec sh -c 'L=`basename "{}" .po` && mkdir -p "$RPM_BUILD_ROOT/usr/share/YaST2/locale/$L/LC_MESSAGES" && msgfmt -o "$RPM_BUILD_ROOT/usr/share/YaST2/locale/$L/LC_MESSAGES/agama.mo" "{}"' \;
+if [ "$1" = "-d" ]; then
+  is_archive=false
+  PODIR="$2"
+else
+  is_archive=true
+  PODIR=$(mktemp --directory --suffix "-agama-po")
+  tar xfjv "$1" -C "$PODIR"
+fi
 
-rm -rf "$PODIR"
+export DESTDIR="$RPM_BUILD_ROOT"
+# export localedir=/usr/share/locale
+export localedir=/usr/share/YaST2/locale
+
+find "$PODIR" -name "*.po" -exec sh -c '
+  LL=`basename "$1" .po` &&
+    mkdir -p  "${DESTDIR}${localedir}/$LL/LC_MESSAGES" &&
+    msgfmt -o "${DESTDIR}${localedir}/$LL/LC_MESSAGES/agama.mo" "$1"
+  ' sh {} \;
+
+if $is_archive; then
+  rm -rf "$PODIR"
+fi
